new scripts and additions in exisiting scripts to select reduced  area
In a new experiment version, the exploration time should be reduced to 30 min only. Therefore a new area of the city needs to be selected, to reduce the walkable area to a reasoneable size for 30 min. The new changes either calculate some statistics in existing scripts to identify some interesting variables taken into account for selecting the new area or directly identify the new area and calculate stats for them.